code refactoring; fixed hard-links URI for confirming to FCGI protocol; auto redirect /path/to/folder/ to /path/to/folder/index if index exists

+++ zbox_wiki/main.py
@@ -4,31 +4,31 @@ import cgi
 import os
 import shutil
 
+# ship web.py with this project for walking around custom static files folder bug
+import web
+
 import acl
 import cache
+import consts
 import commons
 import page
 import config_agent
-import paginator
 import search
 import static_file
-import web
 import mdutils
 
 
+# declare for using in scripts/fcgi_main.py
 __all__ = [
-    "web",
-
-    "app",
-    "mapping",
     "main",
-    "fix_pages_path_symlink",
-
+    "web",
+    "mapping",
     "Robots",
     "SpecialWikiPage",
     "WikiPage",
 ]
 
+web.config.debug = True
 
 mapping = (
     "/robots.txt", "Robots",
@@ -36,10 +36,6 @@ mapping = (
     "/(~[a-zA-Z0-9_\-/.]+)", "SpecialWikiPage",
     ur"/([a-zA-Z0-9_\-/.%s]*)" % commons.CJK_RANGE, "WikiPage",
 )
-
-g_redirect_paths = ("favicon.ico", "robots.txt")
-g_special_paths = ("~all", "~recent", "~search", "~settings", "~stat", "~new")
-g_actions = ("edit", "read", "rename", "delete", "source")
 
 app = web.application(mapping, globals())
 folder_templates_full_path = config_agent.get_full_path("paths", "templates_path")
@@ -66,7 +62,7 @@ You could fork it and commit the changes, then send a pull request to the mainta
 
 <pre><code>%s</code></pre>"""
 
-        # simple wrapper in CSS
+        # simple E-mail wrapper in CSS
         email = maintainer_email.replace("@", " &lt;AT&gt; ")
         buf = ro_tpl_p1 % email
 
@@ -77,50 +73,6 @@ You could fork it and commit the changes, then send a pull request to the mainta
         web.Forbidden.message = buf
 
 
-def wp_view_settings():
-    enable_show_full_path = web.cookies().get("zw_show_full_path", config_agent.config.get("frontend", "enable_show_full_path"))
-    enable_show_full_path = int(enable_show_full_path)
-
-    enable_auto_toc = web.cookies().get("zw_auto_toc", config_agent.config.getboolean("frontend", "enable_auto_toc"))
-    enable_auto_toc = int(enable_auto_toc)
-
-    enable_highlight = web.cookies().get("zw_highlight", config_agent.config.get("frontend", "enable_highlight"))
-    enable_highlight = int(enable_highlight)
-
-    return tpl_render.view_settings(enable_show_full_path = enable_show_full_path,
-                                    enable_auto_toc = enable_auto_toc,
-                                    enable_highlight = enable_highlight,
-                                    static_files = static_file.g_global_static_files)
-
-
-def wp_get_all_pages(enable_show_full_path, limit, offset):
-    buf = cache.get_all_pages_list_from_cache()
-    all_lines = buf.split()
-    total_lines = len(all_lines)
-
-    title = "All Pages List (%d/%d)" % (offset, total_lines / limit)
-
-    start = offset * limit
-    end = start + limit
-    lines = all_lines[start : end]
-
-    buf = mdutils.sequence_to_unorder_list(lines, enable_show_full_path = enable_show_full_path)
-    folder_pages_full_path = config_agent.get_full_path("paths", "pages_path")
-    content = mdutils.md2html(text = buf, work_full_path = folder_pages_full_path)
-
-    pg = paginator.Paginator()
-    pg.total = total_lines
-    pg.current_offset = offset
-    pg.limit = limit
-    pg.url = "/~all"
-
-    return tpl_render.canvas(config = config_agent.config,
-                             button_path = title,
-                             content = content,
-                             static_files = static_file.g_global_static_files,
-                             paginator = pg)
-
-
 class WikiPage(object):
     @acl.check_ip
     @acl.check_rw
@@ -129,33 +81,19 @@ class WikiPage(object):
 
         inputs = web.input()
         action = inputs.get("action", "read")
-        if action not in g_actions:
+        if action not in consts.g_actions:
             raise web.BadRequest()
 
-        show_Full_path_def = config_agent.config.getboolean("frontend", "enable_show_full_path")
-        auto_toc_def = config_agent.config.getboolean("frontend", "enable_auto_toc")
-        highlight_def = config_agent.config.getboolean("frontend", "enable_highlight")
-
-        enable_show_full_path = int(web.cookies().get("zw_show_full_path", show_Full_path_def))
-        enable_auto_toc = int(web.cookies().get("zw_auto_toc", auto_toc_def))
-        enable_highlight = int(web.cookies().get("zw_highlight", highlight_def))
-
         if action == "read":
-            if req_path == "":
-                req_path = "home"
-
-            return page.wp_read(req_path = req_path,
-                                     enable_show_full_path = enable_show_full_path,
-                                     enable_auto_toc = enable_auto_toc,
-                                     enable_highlight = enable_highlight)
+            return page.wp_read(config_agent = config_agent, req_path = req_path, tpl_render = tpl_render)
         elif action == "edit":
-            return page.wp_edit(req_path)
+            return page.wp_edit(config_agent = config_agent, req_path = req_path, tpl_render = tpl_render)
         elif action == "rename":
-            return page.wp_rename(req_path)
+            return page.wp_rename(config_agent = config_agent, req_path = req_path, tpl_render = tpl_render)
         elif action == "delete":
-            return page.wp_delete(req_path)
+            return page.wp_delete(config_agent = config_agent, req_path = req_path, tpl_render = tpl_render)
         elif action == "source":
-            return page.wp_source(req_path)
+            return page.wp_source(config_agent = config_agent, req_path = req_path, tpl_render = tpl_render)
         else:
             raise web.BadRequest()
 
@@ -169,57 +107,23 @@ class WikiPage(object):
         if (not action) or (action not in ("edit", "rename")):
             raise web.BadRequest()
 
-        content = inputs.get("content")
-        content = web.utils.safestr(content)
-
-        # NOTICE: if req_path == `users/`, full_path will be `/path/to/users/`,
-        #         its parent will be `/path/to/users`.
-        full_path = mdutils.req_path_to_full_path(req_path)
-
-        parent = os.path.dirname(full_path)
-        if not os.path.exists(parent):
-            os.makedirs(parent)
+        new_content = inputs.get("content")
+        new_content = web.utils.safestr(new_content)
 
         if action == "edit":
-            page.update_page_by_req_path(req_path = req_path, content = content)
-
-            web.seeother("/%s" % req_path)
-            return
-        elif action == "rename":
-            new_path = inputs.get("new_path")
-            if not new_path:
+            if (req_path in consts.g_special_paths) or (req_path in consts.g_redirect_paths):
                 raise web.BadRequest()
 
-            old_full_path = mdutils.req_path_to_full_path(req_path)
-            if os.path.isfile(old_full_path):
-                new_full_path = mdutils.req_path_to_full_path(new_path)
-            elif os.path.isdir(old_full_path):
-                folder_pages_full_path = config_agent.get_full_path("paths", "pages_path")
-                new_full_path = os.path.join(folder_pages_full_path, new_path)
-            else:
-                raise Exception("un-expected path '%s'" % req_path)
+            page.wp_update(config_agent = config_agent, req_path = req_path, new_content = new_content)
+            return
 
-            if os.path.exists(new_full_path):
-                err_info = "WARNING: The page %s already exists" % new_full_path
-                return tpl_render.rename(req_path, err_info, static_files = static_file.g_global_static_files)
+        elif action == "rename":
+            new_path = inputs.get("new_path")
+            if (req_path in consts.g_special_paths) or (req_path in consts.g_redirect_paths) or (not new_path):
+                raise web.BadRequest()
 
-            parent = os.path.dirname(new_full_path)
-            if not os.path.exists(parent):
-                os.makedirs(parent)
-
-            shutil.move(old_full_path, new_full_path)
-
-            cache.update_all_pages_list_cache()
-            cache.update_recent_change_cache()
-
-            if os.path.isfile(new_full_path):
-                web.seeother("/%s" % new_path)
-                return
-            elif os.path.isdir(new_full_path):
-                web.seeother("/%s/" % new_path)
-                return
-            else:
-                raise Exception("un-expected path '%s'" % new_path)
+            page.wp_rename_post(config_agent = config_agent, tpl_render = tpl_render, req_path = req_path, new_path = new_path)
+            return
 
         url = os.path.join("/", req_path)
         web.redirect(url)
@@ -230,28 +134,29 @@ class SpecialWikiPage(object):
     @acl.check_ip
     @acl.check_rw
     def GET(self, req_path):
-        assert req_path in g_special_paths
-
         inputs = web.input()
+
         FIRST_PAGE = 0
         offset = int(inputs.get("offset", FIRST_PAGE))
 
         page_limit = config_agent.config.getint("pagination", "page_limit")
         limit = int(inputs.get("limit", page_limit))
 
-        enable_show_full_path = config_agent.config.getboolean("frontend", "enable_show_full_path")
-        enable_show_full_path = int(web.cookies().get("zw_show_full_path", enable_show_full_path))
-
         if req_path == "~recent":
-            return cache.wp_get_recent_changes_from_cache(enable_show_full_path = enable_show_full_path, limit = limit, offset = offset)
+            return page.wp_get_recent_changes_from_cache(config_agent = config_agent, tpl_render = tpl_render,
+                                                         req_path = req_path, limit = limit, offset = offset)
         elif req_path == "~all":
-            return wp_get_all_pages(enable_show_full_path = enable_show_full_path, limit = limit, offset = offset)
+            return page.wp_get_all_pages(config_agent = config_agent, tpl_render = tpl_render, req_path = req_path,
+                                         limit = limit, offset = offset)
         elif req_path == "~settings":
-            return wp_view_settings()
+            return page.wp_view_settings(config_agent = config_agent, tpl_render = tpl_render, req_path = req_path)
+
         elif req_path == "~stat":
-            return page.wp_stat()
+            return page.wp_stat(config_agent = config_agent, tpl_render = tpl_render, req_path = req_path)
+
         elif req_path == "~new":
-            return page.wp_new()
+            return page.wp_new(config_agent = config_agent, tpl_render = tpl_render, req_path = req_path)
+
         else:
             return web.BadRequest()
 
@@ -261,52 +166,35 @@ class SpecialWikiPage(object):
         inputs = web.input()
             
         if req_path == "~search":
-            keywords = inputs.get("k")
-            keywords = web.utils.safestr(keywords)
-            if keywords:
-                arg = web.cookies().get("zw_show_full_path") or config_agent.config.get("frontend", "enable_show_full_path")
-                enable_show_full_path = int(arg)
-
-                limit = config_agent.config.getint("pagination", "search_page_limit")
-                lines = search.search_by_filename_and_file_content(keywords, limit = limit)
-                content = mdutils.sequence_to_unorder_list(seq = lines, enable_show_full_path = enable_show_full_path)
-            else:
-                content = None
-
-            if content:
-                content = mdutils.md2html(content)
-            else:
-                content = "matched not found"
-
-            return tpl_render.search(keywords = keywords, content = content, static_files = static_file.g_global_static_files)
+            return page.wp_search(config_agent = config_agent, tpl_render = tpl_render, req_path = req_path)
 
         elif req_path == "~settings":
-            enable_show_full_path = inputs.get("enable_show_full_path")
-            enable_auto_toc = inputs.get("enable_auto_toc")
-            enable_highlight = inputs.get("enable_highlight")
+            show_full_path = inputs.get("show_full_path")
+            auto_toc = inputs.get("auto_toc")
+            highlight_code = inputs.get("highlight_code")
 
-            if enable_show_full_path == "on":
-                enable_show_full_path = 1
+            if show_full_path == "on":
+                show_full_path = 1
             else:
-                enable_show_full_path = 0
-            web.setcookie(name = "zw_show_full_path", value = enable_show_full_path, expires = 31536000)
+                show_full_path = 0
+            web.setcookie(name = "zw_show_full_path", value = show_full_path, expires = 31536000)
 
-            if enable_auto_toc == "on":
-                enable_auto_toc = 1
+            if auto_toc == "on":
+                auto_toc = 1
             else:
-                enable_auto_toc = 0
-            web.setcookie(name = "zw_auto_toc", value = enable_auto_toc, expires = 31536000)
+                auto_toc = 0
+            web.setcookie(name = "zw_auto_toc", value = auto_toc, expires = 31536000)
 
-            if enable_highlight == "on":
-                enable_highlight = 1
+            if highlight_code == "on":
+                highlight_code = 1
             else:
-                enable_highlight = 0
-            web.setcookie(name = "zw_highlight", value = enable_highlight, expires = 31536000)
+                highlight_code = 0
+            web.setcookie(name = "zw_highlight", value = highlight_code, expires = 31536000)
 
 
             latest_req_path = web.cookies().get("zw_latest_req_path")
 
-            if latest_req_path and (latest_req_path not in g_redirect_paths) and latest_req_path != "/":
+            if latest_req_path and (latest_req_path not in consts.g_redirect_paths) and latest_req_path != "/":
                 web.setcookie(name = "zw_latest_req_path", value = "", expires = -1)
                 latest_req_path = "/" + latest_req_path
             else:
@@ -314,20 +202,23 @@ class SpecialWikiPage(object):
 
             web.seeother(latest_req_path)
             return
+
         elif req_path == "~new":
-            real_req_path = inputs.get("path")
-            fixed_req_path = web.lstrips(real_req_path, "/")
+            buf_path = inputs.get("path")
+            buf_path = commons.strutils.lstrips(buf_path, "/")
+            buf_path = commons.strutils.rstrips(buf_path, ".md")
+            fixed_path = commons.strutils.rstrips(buf_path, ".markdown")
+            if (not fixed_path) or (fixed_path in consts.g_special_paths) or (fixed_path in consts.g_redirect_paths):
+                raise web.BadRequest()
 
             content = inputs.get("content")
             content = web.utils.safestr(content)
-        
-            page.update_page_by_req_path(req_path = fixed_req_path, content = content)
+            if not content:
+                raise web.BadRequest()
 
-            cache.update_recent_change_cache()
-            cache.update_all_pages_list_cache()
-
-            web.seeother(real_req_path)
+            page.wp_create(config_agent = config_agent, req_path = req_path, path = fixed_path, content = content)
             return
+
         else:
             raise web.NotFound()
 
@@ -336,7 +227,7 @@ class Robots(object):
     def GET(self):
         folder_pages_full_path = config_agent.get_full_path("paths", "pages_path")
         path = os.path.join(folder_pages_full_path, "robots.txt")
-        content = commons.cat(path)
+        content = commons.shutils.cat(path)
 
         web.header("Content-Type", "text/plain")
         return content
@@ -361,12 +252,12 @@ def fix_pages_path_symlink(proj_root_full_path):
     src_full_path = os.path.join(proj_root_full_path, "pages")
     dst_full_path = os.path.join(proj_root_full_path, "static", "pages")
 
-    if os.path.islink(dst_full_path):
-        os.remove(dst_full_path)
-        
+    if os.path.islink(dst_full_path) and os.readlink(dst_full_path) != src_full_path:
+        if os.path.islink(dst_full_path):
+            os.remove(dst_full_path)
+
     if not os.path.exists(dst_full_path):
         os.symlink(src_full_path, dst_full_path)
-
 
 def main(instance_root_full_path):
     web.config.debug = config_agent.config.getboolean("main", "debug")
