+++ zbox_wiki/scripts/zwadmin.py
@@ -1,18 +1,11 @@
 #!/usr/bin/env python
 import ConfigParser
-import difflib
-import filecmp
-import glob
-import logging
 import os
 import platform
 import shutil
 
 from commons import argparse
 import zbox_wiki
-
-
-logging.getLogger().setLevel(logging.INFO)
 
 
 ZW_MOD_FULL_PATH = zbox_wiki.__path__[0]
@@ -63,10 +56,18 @@ Please report bug to shuge.lee <AT> GMail.
 """
 
 
-parser = argparse.ArgumentParser(description = "create/upgrade ZBox Wiki instance", 
+parser = argparse.ArgumentParser(description = "create/upgrade ZBox Wiki instance",
                                  epilog = "Please report bug to shuge.lee <AT> GMail.")
 parser.add_argument("--create", help = "full path of instance")
 parser.add_argument("--upgrade", help = "full path of instance")
+
+
+def get_ans(msg, expect_ans_list = ("Y", "y", "N", "n", "A", "a")):
+    ans = raw_input(msg)
+    while ans not in expect_ans_list:
+        print "expected answer in", expect_ans_list
+        ans = raw_input(msg)
+    return ans
 
 
 def print_zwd_help_msg(instance_full_path):
@@ -106,6 +107,38 @@ def cp_fcgi_scripts(instance_full_path):
         with open(nginx_conf_path, "w") as f:
             f.write(buf)
 
+def fix_folder_pages_sym_link(instance_full_path):
+    src = os.path.join(instance_full_path, "pages")
+    dst = os.path.join(instance_full_path, "static", "pages")
+
+    if os.path.islink(dst):
+        got = os.readlink(dst)
+        if got != src:
+            msg = "expected %s -> %s, got %s" % (src, dst, got)
+            print msg
+            os.remove(dst)
+
+            msg = "link %s -> %s" % (src, dst)
+            print msg
+            os.symlink(src, dst)
+
+    elif os.path.isdir(dst) and (not os.path.islink(dst)):
+        msg = "expected %s is a symbolic link, got a directory, delete them" % dst
+        print msg
+        shutil.rmtree(dst)
+
+        msg = "link %s -> %s" % (src, dst)
+        print msg
+        os.symlink(src, dst)
+
+    elif os.path.isfile(dst):
+        msg = "expected %s is a symbolic link, got a file, delete it" % dst
+        print msg
+        os.remove(dst)
+
+        msg = "link %s -> %s" % (src, dst)
+        print msg
+        os.symlink(src, dst)
 
 def action_create(instance_full_path):
     for folder_name in ("static", "templates", "pages"):
@@ -113,70 +146,85 @@ def action_create(instance_full_path):
         dst = os.path.join(instance_full_path, folder_name)
 
         if os.path.exists(dst):
-            msg = dst + " already exists, skip \n"
-            logging.warn(msg)
+            msg = "%s already exists, skip" % dst
+            print msg
             continue
-
         shutil.copytree(src, dst)
 
+    fix_folder_pages_sym_link(instance_full_path)
 
     for folder_name in ("tmp", "sessions"):
         src_full_path = os.path.join(instance_full_path, folder_name)
 
         if os.path.exists(src_full_path):
-            msg = src_full_path + "already exists, skip \n"
-            logging.warn(msg)
+            msg = "%s already exists, skip" % src_full_path
+            print msg
             continue
-
         os.mkdir(src_full_path)
 
     src = os.path.join(ZW_MOD_FULL_PATH, "default.cfg")
     dst = os.path.join(instance_full_path, "default.cfg")
-    shutil.copyfile(src, dst)
-    os.chmod(dst, 0754)
+    if os.path.exists(dst):
+        msg = "%s already exists, recover it from default? [Y]es / [N]o " % dst
+        ans = get_ans(msg, expect_ans_list = ["Y", "y", "N", "n"])
+        if ans in ["Y", "y"]:
+            print "copy %s -> %s" % (src, dst)
+            shutil.copyfile(src, dst)
+    else:
+        shutil.copyfile(src, dst)
+    os.chmod(dst, 0644)
 
     cp_fcgi_scripts(instance_full_path)
     print_zwd_help_msg(instance_full_path)
 
 
-def diff_nuar(a_file_path, b_file_path):
-    fromlines = file(a_file_path).readlines()
-    tolines = file(b_file_path).readlines()
-
-    return list(difflib.context_diff(fromlines, tolines, fromfile = a_file_path, tofile = b_file_path))
-
-
 def action_upgrace(instance_full_path):
     folders = ("static", "templates")
-    for i in folders:
-        a = os.path.join(instance_full_path, i)
-        b = os.path.join(ZW_MOD_FULL_PATH, i)
-        cmp_obj = filecmp.dircmp(a, b, hide = [os.curdir, os.pardir] + glob.glob('.*'))
+    yes_to_all = False
 
-        for j in cmp_obj.diff_files:
-            msg = "%s in instance has changed, recover it from default? [Y/n]" % j
-            ans = raw_input(msg)
-            if ans in ("Y", "y"):
-                src = os.path.join(ZW_MOD_FULL_PATH, i, j)
-                dst = os.path.join(instance_full_path, i, j)
-                shutil.copy(src, dst)
+    for i in folders:
+        src = os.path.join(ZW_MOD_FULL_PATH, i)
+        dst = os.path.join(instance_full_path, i)
+
+        if os.path.exists(dst):
+            msg = "%s already exists, recover it from default?  [Y]es / [N]o / yes to [A]ll " % dst
+            if yes_to_all:
+                shutil.rmtree(dst)
+                
+                print "copy %s -> %s" % (src, dst)
+                shutil.copytree(src, dst)
+            else:
+                ans = get_ans(msg)
+                if ans in ["A", "a"]:
+                    yes_to_all = True
+                if ans in ["Y", "y", "A", "a"]:
+                    shutil.rmtree(dst)
+                    
+                    print "copy %s -> %s" % (src, dst)
+                    shutil.copytree(src, dst)
+        else:
+            print "copy %s -> %s" % (src, dst)
+            shutil.copytree(src, dst)
+
+    fix_folder_pages_sym_link(instance_full_path)
 
 
     instance_config_file = os.path.join(instance_full_path, "default.cfg")
     default_config_file = os.path.join(ZW_MOD_FULL_PATH, "default.cfg")
 
     if not os.path.exists(instance_config_file):
-        msg = "%s does not exists, recover it from default"
-        logging.info(msg)
+        print "copy %s -> %s" % (default_config_file, instance_config_file)
         shutil.copy(default_config_file, instance_config_file)
     else:
         try:
-            zbox_wiki.load_config(paths = [instance_config_file])
+            zbox_wiki.config_agent.load_config(paths = [instance_config_file])
         except ConfigParser.ParsingError:
             msg = "parsing %s failed, recover it from default? [Y/n]" % instance_config_file
             ans = raw_input(msg)
             if ans in ("Y", "y"):
                 shutil.copy(default_config_file, instance_config_file)
+
+    print_zwd_help_msg(instance_full_path)
 
 
 if __name__ == "__main__":
